move functions to generate synthetic calcium traces
